requirements.txt
Flask=2.0.2
pika=24.3.1
mysql-connector-python=8.0.26